+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to `audiostack` will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.10.0] - 2025-02-13
+
+### Added
+
+- Customer trace id header can now be set per production using the `use_trace` context manager.
+- Can pass through custom headers into `send_request`.
+
 ## [2.9.0] - 2025-01-23
 
 ### Fixed